requirements.txt
asgiref==3.5.0
cycler==0.11.0
Django==4.0.3
django-cors-headers==3.11.0
fonttools==4.31.1
kiwisolver==1.4.0
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
